// src/shared/services/api/generated/apiComponents.ts
export interface GetPostsQueryParams {
  /**
   * Filter posts by user ID
   */
  userId?: number
}

export type GetPostsError = Fetcher.ErrorWrapper<undefined>

export type GetPostsResponse = Schemas.Post[]

export type GetPostsVariables = {
  queryParams?: GetPostsQueryParams
} & ApiContext['fetcherOptions']
// ...
export interface GetPostPathParams {
  /**
   * Post ID
   */
  id: number
}

export type GetPostError = Fetcher.ErrorWrapper<undefined>

export type GetPostVariables = {
  pathParams: GetPostPathParams
} & ApiContext['fetcherOptions']
// ...
export interface GetPostCommentsPathParams {
  /**
   * Post ID
   */
  postId: number
}

export type GetPostCommentsError = Fetcher.ErrorWrapper<undefined>

export type GetPostCommentsResponse = Schemas.Comment[]

export type GetPostCommentsVariables = {
  pathParams: GetPostCommentsPathParams
} & ApiContext['fetcherOptions']
// ...
export interface GetCommentsQueryParams {
  /**
   * Filter comments by post ID
   */
  postId?: number
}

export type GetCommentsError = Fetcher.ErrorWrapper<undefined>

export type GetCommentsResponse = Schemas.Comment[]

export type GetCommentsVariables = {
  queryParams?: GetCommentsQueryParams
} & ApiContext['fetcherOptions']
// ...
export type GetUsersVariables = ApiContext['fetcherOptions']
// ...
export interface GetUserPathParams {
  /**
   * User ID
   */
  id: number
}

export type GetUserError = Fetcher.ErrorWrapper<undefined>

export type GetUserVariables = {
  pathParams: GetUserPathParams
} & ApiContext['fetcherOptions']
// ...
export interface GetTodosQueryParams {
  /**
   * Filter todos by user ID
   */
  userId?: number
  /**
   * Filter by completion status
   */
  completed?: boolean
}

export type GetTodosError = Fetcher.ErrorWrapper<undefined>

export type GetTodosResponse = Schemas.Todo[]

export type GetTodosVariables = {
  queryParams?: GetTodosQueryParams
} & ApiContext['fetcherOptions']
// ...
export interface GetTemplatesQueryParams {
  /**
   * Search by name
   */
  search?: string
  /**
   * Filter by status
   */
  status?: 'active' | 'inactive' | 'draft'
  /**
   * Page number
   *
   * @default 1
   */
  page?: number
  /**
   * Items per page
   *
   * @default 10
   */
  limit?: number
}

export type GetTemplatesError = Fetcher.ErrorWrapper<undefined>

export type GetTemplatesVariables = {
  queryParams?: GetTemplatesQueryParams
} & ApiContext['fetcherOptions']
// ...
export interface GetTemplatePathParams {
  /**
   * Template ID
   *
   * @format uuid
   */
  id: string
}

export type GetTemplateError = Fetcher.ErrorWrapper<undefined>

export type GetTemplateVariables = {
  pathParams: GetTemplatePathParams
} & ApiContext['fetcherOptions']
// ...
export type QueryOperation =
  | {
      path: '/posts'
      operationId: 'getPosts'
      variables: GetPostsVariables | reactQuery.SkipToken
    }
  | {
      path: '/posts/{id}'
      operationId: 'getPost'
      variables: GetPostVariables | reactQuery.SkipToken
    }
  | {
      path: '/posts/{postId}/comments'
      operationId: 'getPostComments'
      variables: GetPostCommentsVariables | reactQuery.SkipToken
    }
  | {
      path: '/comments'
      operationId: 'getComments'
      variables: GetCommentsVariables | reactQuery.SkipToken
    }
  | {
      path: '/users'
      operationId: 'getUsers'
      variables: GetUsersVariables | reactQuery.SkipToken
    }
  | {
      path: '/users/{id}'
      operationId: 'getUser'
      variables: GetUserVariables | reactQuery.SkipToken
    }
  | {
      path: '/todos'
      operationId: 'getTodos'
      variables: GetTodosVariables | reactQuery.SkipToken
    }
  | {
      path: '/templates'
      operationId: 'getTemplates'
      variables: GetTemplatesVariables | reactQuery.SkipToken
    }
  | {
      path: '/templates/{id}'
      operationId: 'getTemplate'
      variables: GetTemplateVariables | reactQuery.SkipToken
    }